Move Logger type to LogHandler
